refactor: moved forms' state related logic to parent pages
