+++ frontend/src/pages/CreateRecipe.tsx
@@ -1,4 +1,15 @@
+import { useNavigate } from "react-router-dom";
 import Form from "../components/Form";
+import recipeService from "../services/recipes";
+
 export default function CreateRecipe() {
-  return <Form />;
+  const navigate = useNavigate();
+
+  const onSubmit = async (recipeObject: RecipeType) => {
+    const returnedObject = await recipeService.create(recipeObject);
+    navigate(`/recipes/${returnedObject._id}`);
+  };
+
+  // onsubmit component
+  return <Form onSubmit={onSubmit} />;
 }
